Added fragments and semi-functional navigation bar
